requirements.txt
PyMySQL==1.1.0
PyQt5_Frameless_Window==0.3.9
websockets==12.0
